Finish raw version

+++ src/lta/cmdi_processes.py
@@ -1,50 +1,44 @@
 import os
-import subprocess
 import shutil
 import xml.etree.ElementTree as ET
-from media_files import check_directory_integrity
+from lxml import etree
+from pathlib import Path
 
-import lta.session_cmdi as cmdi
+from lta.mediatypes import is_media_file, is_transcript_file, mediatype_from_filename
+from lta.media_files import check_directory_integrity
+from lta.xml_validation import validate_session_cmdi
+from lta.session_cmdi import change_resource_proxy_list, change_media_session_bundle, change_written_resources
 from lta.files import create_directory_if_not_exists
 
 
 ET.register_namespace('', 'http://www.clarin.eu/cmd/')
 
 
-def process_session_cmdi(input_file, output_file, media_dir):
+def pretty_print_lxml(input):
+    temp = etree.fromstring(input)
+    pretty = etree.tostring(temp, pretty_print=True)
+    return pretty
+
+
+def process_session_cmdi(input_file, output_file, media_dir, dry_run):
     """Check media files and enricht original session cmdi with media data."""
 
-    print(f'Processing file {input_file}...')
     # Directory scanning.
-
     base_path = media_dir
-    stem = input_file.stem
+    stem = Path(input_file).stem
     interview_id = stem
-    dir_path = base_path # os.path.join(base_path, stem)
+    dir_path = base_path
     files = os.listdir(dir_path)
 
-    def filter_media_files(filename):
-        pair = os.path.splitext(filename)
-        ext = pair[1].lower()
-        return ext in ['.m2ts', '.mp4', '.avi']
-
-    def filter_transcript_files(filename):
-        pair = os.path.splitext(filename)
-        ext = pair[1].lower()
-        return ext in ['.ods', '.pdf', '.csv']
-
-
-    media_files = list(filter(filter_media_files, list(files)))
-    transcript_files = list(filter(filter_transcript_files, list(files)))
-
+    media_files = list(filter(is_media_file, list(files)))
     media_files.sort()
+
+    transcript_files = list(filter(is_transcript_file, list(files)))
     transcript_files.sort()
 
     # Get media_type from first media file.
     first_media_file = media_files[0]
-    print(first_media_file)
-    media_type = cmdi.get_media_type(first_media_file)
-
+    media_type = mediatype_from_filename(first_media_file)
 
     if (len(media_files) != len(transcript_files)):
         raise ValueError(f'{input_file}: Number of media files does not match number of transcript files.')
@@ -59,42 +53,28 @@ def process_session_cmdi(input_file, output_file, media_dir):
     tree = ET.parse(input_file)
     root = tree.getroot()
 
-    cmdi.change_resource_proxy_list(root, interview_id, media_files, transcript_files)
-    cmdi.change_media_session_bundle(root, num_parts, media_type, transcript_files)
-    cmdi.change_written_resources(root)
+    change_resource_proxy_list(root, interview_id, media_files, transcript_files)
+    change_media_session_bundle(root, num_parts, media_type, transcript_files)
+    change_written_resources(root)
 
-    actors = cmdi.get_actors(root)
 
     # Write to tempfile
-    output_str = ET.tostring(root, encoding='utf-8', xml_declaration=True)
-    #tree.write(output_file, encoding='utf-8', xml_declaration=True)
+    output = ET.tostring(root, encoding='utf-8', xml_declaration=True)
+    prettified_output = pretty_print_lxml(output)
 
-    # Subprocess pretty
-    cp1 = subprocess.run(['xmllint', '--format', '-'],
-        input=output_str, capture_output=True)
-
-    if cp1.returncode == 0:
-        print('Prettified…')
-
-    prettified_output = cp1.stdout
-
-    # Check integrity.
-    cp2 = subprocess.run(['xmllint', '--schema', 'media-session-profile.xsd',
-        '--noout', '-'],
-        input=prettified_output, capture_output=True)
-
-    if cp2.returncode == 0:
-        print('Validated…')
-    else:
-        print('Not validated', cp2.stderr)
 
     # Save it.
-    with open(output_file, 'wb') as binary_file:
-        binary_file.write(prettified_output)
-        print('Saved…')
+    if dry_run:
+        print(f'[DRYRUN] Saved session cmdi file to {output_file}')
+    else:
+        with open(output_file, 'wb') as f:
+            f.write(prettified_output)
+
+        validate_session_cmdi(output_file)
+        print(f'Saved session cmdi file to {output_file}')
 
 
-def copy_corpus_cmdi(input_dir, output_dir):
+def copy_corpus_cmdi(input_dir, output_dir, dry_run = False):
     """Find the corpus cmdi file in input_dir and copy it to output_dir."""
 
     files = os.listdir(input_dir)
@@ -103,10 +83,15 @@ def copy_corpus_cmdi(input_dir, output_dir):
         filepath = os.path.join(input_dir, file)
 
         if os.path.isfile(filepath):  # corpus cmdi file
-            shutil.copy(filepath, output_dir)
+            if dry_run:
+                print(f'[DRYRUN] Copied corpus cmdi file {filepath} to {output_dir}')
+            else:
+                print(filepath, output_dir)
+                shutil.copy(filepath, output_dir)
+                print(f'Copied corpus cmdi file {filepath} to {output_dir}')
 
 
-def process_session_cmdi_dir(input_dir, output_dir, media_dir):
+def process_session_cmdi_dir(input_dir, output_dir, media_dir, dry_run = False):
     """Find each session cmdi and call process function on it."""
 
     files = os.listdir(input_dir)
@@ -117,9 +102,13 @@ def process_session_cmdi_dir(input_dir, output_dir, media_dir):
         if os.path.isdir(filepath):  # directory with session cmdi.
             input_session_cmdi = os.path.join(input_dir, file, f'{file}.xml')
             interview_output_dir = os.path.join(output_dir, file)
-            output_session_cmdi = os.path.join(output_dir, file, f'{file}.xml')
+            output_session_cmdi = os.path.join(interview_output_dir, f'{file}.xml')
             interview_media_dir = os.path.join(media_dir, file)
 
-            create_directory_if_not_exists(interview_output_dir)
+            if dry_run:
+                print(f'[DRYRUN] Creating directory {interview_output_dir}')
+            else:
+                create_directory_if_not_exists(interview_output_dir)
 
-            process_session_cmdi(input_session_cmdi, output_session_cmdi, interview_media_dir)
+            process_session_cmdi(input_session_cmdi, output_session_cmdi, interview_media_dir,
+                dry_run)
